Replace npm/bower with uncss'ed CSS and its script
So the script could possibly do with some comments.

I touch the uncss'ed file, so that it exists when uncss parses the
template, because if uncss can't find it, it breaks. I tried using the
`--ignoreSheets` command to ignore that stylesheet, but it doesn't do
what I want.

I copied `bootstrap.css` to a simpler place, so the paths were less
messy. I left bower installing to default bower_components so we have
less stuff in the repo.